server/requirements.txt
bcrypt==3.1.7
Flask==1.1.1
Flask-JWT==0.3.2
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
marshmallow-sqlalchemy==0.19.0
mysqlclient==1.4.4
passlib==1.7.1
python-dotenv==0.10.3
